+++ RELEASE_NOTES.md
@@ -1,5 +1,8 @@
 ## [Undefined] 
 
+### Added
+- Multiple verbosity levels allowed on `AnsiblePlaybookOptions`: new attributes are `VerboseV`, `VerboseVV`,`VerboseVVV`, and `VerboseVVVV`
+- Multiple verbosity levels allowed on `AnsibleAdhocOptions`: new attributes are `VerboseV`, `VerboseVV`,`VerboseVVV`, and `VerboseVVVV`
+
 ### Changed
 - bump to go 1.19
-